chore: optimize userId alert
